checkout page

+++ cart.php
@@ -108,13 +108,13 @@ if(isset($_GET['delete_all'])){
     </div>
 
     <div class="dlt">
-        <a href="cart.php?delete_all" class="btn2">delete all</a>
+        <a href="cart.php?delete_all" class="btn2" onclick="return confirm('do you want ot delete all from cart ?')">delete all</a>
     </div>
 
     <div class="wishlist_total">
         <p>total amount payable : <span>$<?php echo $grand_total?>/=</span></p>
         <a href="shop.php" class="btn2"> continue shopping</a>
-        <a href="checkout.php" class="btn2 <?php echo($grand_total > 1)?'':'disabled'?>" onclick="return confirm('do you want ot delete all from cart ?')">proceed to check out</a>
+        <a href="checkout.php" class="btn2 <?php echo($grand_total > 1)?'':'disabled'?>" >proceed to check out</a>
     </div>
     
     </div>
